+++ CHANGELOG.md
@@ -25,6 +25,8 @@
 * Fix Edit in quickMARC and View Source options state after instance editing. Refs UIIN-1204.
 * Change request status for `Awaiting delivery` items marked as missing. Fixes UIIN-1206.
 * In transit report does not export name of library. Refs UIIN-1058.
+* `Holdings - Call number, eye readable`. Add search only by the call number. Refs UIIN-1040.
+* Replace `bigtest/mirage` with `miragejs`.
 
 ## [4.0.1](https://github.com/folio-org/ui-inventory/tree/v4.0.1) (2020-07-01)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v4.0.0...v4.0.1)